requirements.txt
requests==2.25.1
urllib3==1.26.5
flask
flask-restplus
python-keycloak